backend/pyproject.toml
[tool.poetry]
name = "synergy-link-backend"
version = "0.1.0"
description = "Backend API for Synergy Link Personal Vision Map Generator"
authors = ["Devin AI <devin-ai-integration[bot]@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
bcrypt = "^4.3.0"
email-validator = "^2.2.0"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
